If the single container root URL is a folder then do an extra check
When the single root URL is a folder we need to make sure
the resource .class file actually exists under the root directory.
This is because loose configurations may map a directory to a path
of a package making the content come from a non-root container.
We need to handle this case where there may be one root container
folder and one additional folder mapped to a sub folder as a non-root
container. In this case resources from the sub folder may come from
a directory on disk that is not a sub folder of the root folder URL.