Use a range with repo macro for SR OpenAPI modules
When including resources using the ${repo} macro, bnd will treat a
version as a lower bound with no upper bound and choose the highest
matching version to satisfy the dependency.

Since we don't want any new minor version to be picked up by these
projects, we need to specify either a range with an upper bound or use
the EXACT strategy.