Add tests for patch libraries
This tests that the order is correct for searching various patch libraries
and common libraries.  The test application contains 8 classes that
are also provided by the various libraries configured as patch or
common.  It confirms that the correct patch library can override the class
contained in the application and that no common library is able to
override the class contained in the application.

The assertion is done by checking the codesource location is using
the expected JAR or WAR file.

Additional assertions are done for resource lookups to validate
the correct resource is used for getResource and that the order
is correct when calling getReources.